Decided that a result just can't be a uint32_t
Because of messiness in conversions between float and int.  Sorry, sampno, you'll just roll over at 32768.  That's enough for over a year at 5 minutes anyway.

+++ src/sensors/ProcessorStats.h
@@ -56,7 +56,7 @@ public:
 private:
     const char *_version;
     int8_t _batteryPin;
-    uint32_t sampNum;
+    int16_t sampNum;
 };
